Refactored and Optimized (Mostly Stable)
Refactored the GeneticSimulator class to use multi-threading to improve
the runtime of the algorithm. This is almost entirely stable, however,
occasional exceptions arise near the end of the simulation which leads
me to believe there may be a delay to the closing of the worker threads.
Also performed some minor refactoring.