Updated Saving and Loading to Binary Format
Mainly updated the save and load feature to save a board along with the
parameters used all within a binary file format. This also invovled
adding a timer to the simulator so that this information could be stored
in the file. Also, preformed some minor refactoring as I came across
things, such as updating the stack trace messages to only appear when a
debug mode is turned on. To support saving and loading, also added code
for error message boxes and cleaned up the style of the menus.